refactor: drop --tui flag, consolidate TUI tests, move screenshots out of docs
- Remove the `docetl run --tui/--no-tui` command-line option and the
  `_force_tui` plumbing. The progress view is enabled solely by the
  `pipeline.interactive_ui` flag in the config, as intended.
- Consolidate the TUI tests: the three files (tracker, profiles, integration)
  are now one `tests/test_tui.py`, and the three screenshot scripts are now one
  `tests/tui_screenshots.py` that captures both the real-run shots and the
  synthetic large-run heatmap.
- Move the screenshots out of the published docs folder into
  `.github/screenshots/`, and remove the internal design write-up from the repo
  (the pull request description covers the design). Update the few code comments
  that pointed at the removed file.
