+++ .gitignore
@@ -81,8 +81,8 @@ node_modules/*
 reaction_index/
 *experiments/reasoning/othersystems/biodex/.chroma-biodex/
 
-# Progress-visualization TUI (issue #487): committed demo fixtures + real screenshots
+# Progress view (issue #487): committed demo fixtures + screenshots
 !tests/tui_demo/*.yaml
 !tests/tui_demo/*.json
 tests/tui_demo/*_output.json
-!docs/design/screenshots/*.png
+!.github/screenshots/*.png
